refactor(launcher): extract game-launch side effects into a service
commands/launcher.rs drops from 1031 to 512 lines. Per Clean Architecture
(all fs/process/window/network side effects belong in services/), moved into
new services/game_launch_service.rs:

- launch_with_lr (CreateProcessW spawn via Locale Remulator)
- kill_patcher_loop + find_and_kill_patcher (Toolhelp32 process kill)
- skip_play_window_loop (FindWindowW/PostMessageW)
- find_process_pid_by_name (Toolhelp32 enumeration)
- get_exe_version (PE version query) + get_server_version (TCP handshake)

Command handlers now just delegate. Pure mechanical move — no behavior change.
NOTE: Win32/process code is only truly verifiable via a live game launch;
smoke-test launch + patcher-kill + play-window-skip before merging.
